requirements.txt
Flask
WeasyPrint
gunicorn
psycopg2-binary
openpyxl
pypdf
PyMuPDF
pillow
pytesseract
Flask-Login
python-dotenv